app_backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
idna==3.10
pydub==0.25.1
requests==2.32.3
SpeechRecognition==3.14.2
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
